pyproject.toml
[project]
name = "ai-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
